chore: refactor validator
