+++ src/validator/constraint.ts
@@ -1,7 +1,13 @@
-import { Violation } from '@/validator/violation';
-import { FieldRules } from '@/validator/types';
+import { Violation } from './violation';
 
-export type ConstraintOptions = { groups?: string[] } & Record<string, unknown>;
+export type ConstraintOptions = {
+  groups?: string[];
+  [key: string]: unknown;
+};
+
+type FieldRuleEntry = string | Record<string, ConstraintOptions>;
+export type FieldRules = Record<string, ConstraintOptions> | FieldRuleEntry[];
+
 export type ConstraintContext = {
   path: string;
   root: unknown;
