+++ src/validator/constraints/other/compound.ts
@@ -1,6 +1,5 @@
-import type { FieldRules } from '../../types';
 import { Violation } from '@/validator/violation';
-import { ConstraintContext } from '@/validator/constraint';
+import { ConstraintContext, FieldRules } from '@/validator/constraint';
 
 export function compound(rules: FieldRules) {
   return function compoundConstraint(value: unknown, context: ConstraintContext): Violation[] {
