+++ src/validator/middleware/create-validation-middleware.ts
@@ -1,7 +1,7 @@
 import { type HttpMiddleware, HttpScope, NextMiddleware } from '@/Http';
-import { type ValidationRules, type Validator } from '@/validator/types';
 import { flattenViolations } from '@/validator/flatten-violations';
 import { Violation } from '@/validator/violation';
+import { ValidationRules, Validator } from '@/validator/validator';
 
 type ViolationMapper = (violations: Violation[]) => Record<string, string[]>;
