--- src/validator/options.ts
@@ -1,1 +1,0 @@
-export type Options = { groups?: string[] };
